From: Katsuhiro Kondou <kondou@nec.co.jp>
expire/makehistory.c:
        - QIOclose() should be called before return
frontends/encode.c:
        - cast needed since argument type is different
backends/nntpget.c:
        - DISPOSE unallocated space
nnrpd/article.c:
        - cast position changed
nnrpd/misc.c:
        - function type is void
        - argument type should be time_t
nnrpd/nnrpd.c:
        - cast needed since argument type is different
nnrpd/post.c:
        - cast needed since argument type is different
lib/getfqdn.c:
        - clibrary.h should be included after netdb.h since INADDR_NONE
          may be redefined
storage/interface.c:
        - SMnext() might refer illegal method_data
        - return FALSE if SMcancel() fails
storage/overview.c:
        - cast position changed
        - remove meaningless comparison

+++ frontends/encode.c
@@ -117,7 +117,7 @@ main(ac, av)
 	    p = b3;
 	}
     }
-    encode(b3, p - b3);
+    encode(b3, (int)(p - b3));
     flushout();
     exit(0);
     /* NOTREACHED */
